commit first

+++ database/migrations/1492943785711_create_stock_table.js
@@ -9,7 +9,7 @@ class StocksTableSchema extends Schema {
       table.increments()
       table.timestamps()
       table.string('name',100).unique().notNullable()
-      table.float('amount').notNullable()
+      table.integer('amount').notNullable()
     })
   }
